pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newchan"
version = "0.1.0"
description = "缠论量化分析工具"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "pandas",
    "pyarrow",
    "matplotlib",
    "python-dotenv",
    "ib_insync",
    "lightweight-charts",
    "databento",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
]

[project.scripts]
newchan = "newchan.cli:main"

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools.packages.find]
where = ["src"]
